refactor: move user controller to usersController
